:art: Making absolute path. Preparing package for using in Django
