--- CrawlerEngine/Crawler/Downloaders/__init__.py
@@ -1,10 +1,0 @@
-from .IDownloader import IDownloader
-
-from .DownloaderRequests import DownloaderRequests
-from .DownloaderSelenium import DownloaderSelenium
-
-__all__ = (
-    IDownloader,
-    DownloaderRequests,
-    DownloaderSelenium
-)
