--- CrawlerEngine/Crawler/Parsers/__init__.py
@@ -1,10 +1,0 @@
-from .IParser import IParser
-
-from .ParserJson import ParserJson
-from .ParserXml  import ParserXml
-
-__all__ = (
-    IParser,
-    ParserJson,
-    ParserXml
-)
