--- CrawlerEngine/Crawler/Sources/__init__.py
@@ -1,7 +1,0 @@
-from .Source import Source
-from .SourceLoader import SourceLoader
-
-__all__ = (
-    SourceLoader,
-    Source
-)
