--- CrawlerEngine/Crawler/__init__.py
@@ -1,7 +1,0 @@
-from .Crawler import Crawler
-from .Sources import SourceLoader
-
-__all__ = (
-    SourceLoader,
-    Crawler
-)
