--- CrawlerEngine/__init__.py
@@ -1,3 +1,0 @@
-from .Manager import Manager
-
-__all__ = Manager
