+++ crawler/__init__.py
@@ -1,0 +1,7 @@
+from crawler.manager import (
+    CrawlerManager
+)
+
+__all__ = (
+    CrawlerManager
+)
