+++ crawler/crawler.py
@@ -4,7 +4,9 @@ from typing import Optional
 from concurrent.futures import ThreadPoolExecutor
 from queue import Queue, Empty
 
-from .Sources import Source
+from crawler.sources import (
+    Source
+)
 
 log = logging.getLogger(__name__)
