+++ crawler/downloaders/__init__.py
@@ -1,0 +1,17 @@
+from crawler.downloaders.idownloader import (
+    IDownloader
+)
+
+from crawler.downloaders.downloader_requests import (
+    DownloaderRequests
+)
+
+from crawler.downloaders.downloader_selenium import (
+    DownloaderSelenium
+)
+
+__all__ = (
+    IDownloader,
+    DownloaderRequests,
+    DownloaderSelenium
+)
