+++ crawler/downloaders/downloader_requests.py
@@ -3,8 +3,13 @@ from typing import Optional
 
 import requests
 
-from .IDownloader import IDownloader
-from .HttpHeaders import Headers
+from crawler.downloaders.idownloader import (
+    IDownloader
+)
+
+from crawler.downloaders.http_headers import (
+    Headers
+)
 
 log = logging.getLogger(__name__)
