+++ crawler/downloaders/downloader_selenium.py
@@ -1,5 +1,5 @@
 from typing import Optional
-from .IDownloader import IDownloader
+from .idownloader import IDownloader
 
 class DownloaderSelenium(IDownloader):
     def __init__(self):
