+++ crawler/manager.py
@@ -1,18 +1,21 @@
 import logging
 from typing import Optional
 
-from CrawlerEngine.Crawler import (
-    Crawler,
+from crawler.sources import (
     SourceLoader
 )
 
-from CrawlerEngine.Exceptions import (
+from crawler.crawler import (
+    Crawler
+)
+
+from crawler.exceptions import (
     SourceRulesNotLoaded
 )
 
 log = logging.getLogger(__name__)
 
-class Manager():
+class CrawlerManager():
     def __init__(
         self
     ) -> None:
