+++ crawler/parsers/__init__.py
@@ -1,0 +1,17 @@
+from crawler.parsers.iparser import (
+    IParser
+)
+
+from crawler.parsers.parser_json import (
+    ParserJson
+)
+
+from crawler.parsers.parser_xml import (
+    ParserXml
+)
+
+__all__ = (
+    IParser,
+    ParserJson,
+    ParserXml
+)
