+++ crawler/parsers/parser_json.py
@@ -1,16 +1,21 @@
 import logging
 
 from typing import Optional
-from xmltodict import parse as XMLParse
+from json import loads, JSONDecodeError
 
-from .IParser import IParser
-from .Utils import search, extract
+from crawler.parsers.iparser import (
+    IParser
+)
+
+from crawler.parsers.utils import (
+    search, extract
+)
 
 log = logging.getLogger(__name__)
 
-class ParserXml(IParser):
+class ParserJson(IParser):
     """An inheritor of the `IParser`, which provides the 
-    ability to parse given XML to the list of the requested 
+    ability to parse given JSON to the list of the requested 
     fields.
 
     :param root: root of the data dictionary. From this point
@@ -25,13 +30,13 @@ class ParserXml(IParser):
     """
     def __init__(
         self,
-        root: str,
         fields: dict,
+        root: str = None,
         separator: str = "."
     ) -> None:
         """Constructor method."""
-        self._root = root
         self._fields = fields
+        self._root = root
         self._separator = separator
 
     def parse(
@@ -39,21 +44,21 @@ class ParserXml(IParser):
         data: str
     ) -> Optional[dict]:    
         """This function is the main function of this `Parser`. And, 
-        as you can guess, it performs the parsing of XML content.
+        as you can guess, it performs the parsing of JSON content.
         
         :param data: string with the data to be searched for
         :type data: str
-
+                
         :return: data dictionary from the given paths
-        :rtype: dict if success, else NoneType
+        :rtype: dict if success, else None
         """
 
         try:
-            data_dict = XMLParse(data)
+            data_dict = loads(data)
 
-        except Exception as decodeError:
+        except JSONDecodeError as decodeError:
             log.error(
-                f"XMLDecodeError: Failed to decode the data into XML. {decodeError}"
+                f"JSONDecodeError: Failed to decode the data into JSON. {decodeError}"
             )
             return
