+++ crawler/parsers/parser_xml.py
@@ -1,16 +1,21 @@
 import logging
 
 from typing import Optional
-from json import loads, JSONDecodeError
+from xmltodict import parse as XMLParse
 
-from .IParser import IParser
-from .Utils import search, extract
+from crawler.parsers.iparser import (
+    IParser
+)
+
+from crawler.parsers.utils import (
+    search, extract
+)
 
 log = logging.getLogger(__name__)
 
-class ParserJson(IParser):
+class ParserXml(IParser):
     """An inheritor of the `IParser`, which provides the 
-    ability to parse given JSON to the list of the requested 
+    ability to parse given XML to the list of the requested 
     fields.
 
     :param root: root of the data dictionary. From this point
@@ -25,13 +30,13 @@ class ParserJson(IParser):
     """
     def __init__(
         self,
+        root: str,
         fields: dict,
-        root: str = None,
         separator: str = "."
     ) -> None:
         """Constructor method."""
-        self._fields = fields
         self._root = root
+        self._fields = fields
         self._separator = separator
 
     def parse(
@@ -39,21 +44,21 @@ class ParserJson(IParser):
         data: str
     ) -> Optional[dict]:    
         """This function is the main function of this `Parser`. And, 
-        as you can guess, it performs the parsing of JSON content.
+        as you can guess, it performs the parsing of XML content.
         
         :param data: string with the data to be searched for
         :type data: str
-                
+
         :return: data dictionary from the given paths
-        :rtype: dict if success, else None
+        :rtype: dict if success, else NoneType
         """
 
         try:
-            data_dict = loads(data)
+            data_dict = XMLParse(data)
 
-        except JSONDecodeError as decodeError:
+        except Exception as decodeError:
             log.error(
-                f"JSONDecodeError: Failed to decode the data into JSON. {decodeError}"
+                f"XMLDecodeError: Failed to decode the data into XML. {decodeError}"
             )
             return
