+++ crawler/sources/__init__.py
@@ -1,0 +1,12 @@
+from crawler.sources.source_loader import (
+    SourceLoader
+)
+
+from crawler.sources.source import (
+    Source
+)
+
+__all__ = (
+    SourceLoader,
+    Source
+)
