+++ crawler/sources/source.py
@@ -1,8 +1,13 @@
 from typing import Optional
 from dataclasses import dataclass
 
-from ..Downloaders import IDownloader
-from ..Parsers import IParser
+from crawler.downloaders import (
+    IDownloader
+)
+
+from crawler.parsers import (
+    IParser
+)
 
 @dataclass
 class Source:
