+++ crawler/sources/source_loader.py
@@ -8,20 +8,23 @@ from yaml import (
     YAMLError
 )
 
-from .Source import Source
-from .Models import (
+from crawler.sources.source import (
+    Source
+)
+
+from crawler.sources.rules_models import (
     ParserModel,
     DownloaderModel,
     SourceModel
 )
 
-from ..Downloaders import (
+from crawler.downloaders import (
     IDownloader,
     DownloaderRequests,
     DownloaderSelenium
 )
 
-from ..Parsers import (
+from crawler.parsers import (
     IParser,
     ParserJson,
     ParserXml,
