+++ tests/test_parser_utils.py
@@ -1,6 +1,6 @@
 import unittest
 
-from CrawlerEngine.Crawler.Parsers.Utils import (
+from crawler.parsers.utils import (
     search,
     extract
 )
